refactor the authentication backend setting
The setting is now configured during the `nominate` app startup, which
retrieves it from the configured convention

Additionally, environment configuration is now moved into the convention
configuration.
