+++ nomnom/convention.py
@@ -5,11 +5,13 @@
 from collections.abc import Iterable
 from dataclasses import dataclass
 from datetime import date, datetime
+from pathlib import Path
 from typing import TypedDict
 from urllib.parse import urlparse
 
 from django.http import HttpRequest
 from django.utils.timezone import is_aware, make_aware
+from environ import bool_var, config, group, to_config, var
 
 
 class ConventionException(Exception):
@@ -18,6 +20,81 @@ class ConventionException(Exception):
 
 class ConfigurationError(ConventionException):
     ...
+
+
+BASE_DIR = Path(__file__).resolve().parent.parent
+
+
+def comma_separated_string(env_val: str) -> list[str]:
+    return [v.strip() for v in env_val.strip().split(",") if v.strip()]
+
+
+@config(prefix="NOM")
+class SystemConfiguration:
+    convention_app = var(default=None)
+
+    @config
+    class DB:
+        name = var()
+        host = var()
+        port = var(5432, converter=int)
+        user = var()
+        password = var()
+
+    @config
+    class REDIS:
+        host = var()
+        port = var(6379, converter=int)
+
+    @config
+    class EMAIL:
+        host = var()
+        port = var(587, converter=int)
+        host_user = var(default=None)
+        host_password = var(default=None)
+        use_tls = bool_var(default=True)
+
+    @config
+    class CONVENTION:
+        hugo_packet = var(default=False)
+
+    @config
+    class SENTRY_SDK:
+        dsn = var(default=None)
+        environment = var(default="production")
+
+    debug = bool_var(default=False)
+    sentry_sdk = group(SENTRY_SDK)
+    db = group(DB)
+    redis = group(REDIS)
+    email = group(EMAIL)
+
+    @config
+    class OAUTH:
+        key = var()
+        secret = var()
+        backend = var("nomnom.social_core.ClydeOAuth2")
+
+    @config
+    class LOGGING:
+        oauth_debug = bool_var(False)
+
+    oauth = group(OAUTH)
+
+    secret_key = var()
+
+    static_file_root = var(BASE_DIR / "staticfiles")
+
+    allowed_hosts: list[str] = var("", converter=comma_separated_string)
+
+    allow_username_login: bool = bool_var(False)
+
+    convention = group(CONVENTION)
+
+    logging = group(LOGGING)
+
+
+system_configuration = to_config(SystemConfiguration)
 
 
 class URLSetting(TypedDict):
@@ -76,6 +153,7 @@ class ConventionConfiguration:
     hugo_help_email: str
     hugo_admin_email: str
     registration_email: str
+    authentication_backends: list[str]
     logo: str = "images/logo_blue.png"
     logo_alt_text: str = "NomNom Logo"
     nomination_eligibility_cutoff: date | datetime | None = None
