+++ nomnom/settings.py
@@ -11,22 +11,14 @@ https://docs.djangoproject.com/en/5.0/ref/settings/
 """
 
 import os
-from pathlib import Path
 
 import bleach.sanitizer
 from django.utils.translation import gettext_lazy as _
-from environ import bool_var, config, group, to_config, var
-from icecream import install
+from icecream import install as install_icecream
 
-install()
+from nomnom.convention import system_configuration as cfg
 
-# Build paths inside the project like this: BASE_DIR / 'subdir'.
-BASE_DIR = Path(__file__).resolve().parent.parent
-
-
-def comma_separated_string(env_val: str) -> list[str]:
-    return [v.strip() for v in env_val.strip().split(",") if v.strip()]
-
+install_icecream()
 
 try:
     import debug_toolbar  # noqa
@@ -36,74 +28,6 @@ try:
 except ImportError:
     debug_toolbar_app = None
     debug_toolbar_middleware = None
-
-
-@config(prefix="NOM")
-class AppConfig:
-    convention_app = var(default=None)
-
-    @config
-    class DB:
-        name = var()
-        host = var()
-        port = var(5432, converter=int)
-        user = var()
-        password = var()
-
-    @config
-    class REDIS:
-        host = var()
-        port = var(6379, converter=int)
-
-    @config
-    class EMAIL:
-        host = var()
-        port = var(587, converter=int)
-        host_user = var(default=None)
-        host_password = var(default=None)
-        use_tls = bool_var(default=True)
-
-    @config
-    class CONVENTION:
-        hugo_packet = var(default=False)
-
-    @config
-    class SENTRY_SDK:
-        dsn = var(default=None)
-        environment = var(default="production")
-
-    debug = bool_var(default=False)
-    sentry_sdk = group(SENTRY_SDK)
-    db = group(DB)
-    redis = group(REDIS)
-    email = group(EMAIL)
-
-    @config
-    class OAUTH:
-        key = var()
-        secret = var()
-        backend = var("nomnom.social_core.ClydeOAuth2")
-
-    @config
-    class LOGGING:
-        oauth_debug = bool_var(False)
-
-    oauth = group(OAUTH)
-
-    secret_key = var()
-
-    static_file_root = var(BASE_DIR / "staticfiles")
-
-    allowed_hosts: list[str] = var("", converter=comma_separated_string)
-
-    allow_username_login: bool = bool_var(False)
-
-    convention = group(CONVENTION)
-
-    logging = group(LOGGING)
-
-
-cfg = to_config(AppConfig)
 
 
 # Quick-start development settings - unsuitable for production
@@ -196,7 +120,9 @@ NOMNOM_ALLOW_USERNAME_LOGIN_FOR_MEMBERS = cfg.allow_username_login
 NOMNOM_HUGO_NOMINATION_COUNT = 5
 
 AUTHENTICATION_BACKENDS = [
-    cfg.oauth.backend,
+    # NOTE: the nominate.apps.AppConfig.ready() hook will install handlers in this, as the first
+    # set. Any handler in here will be superseded by those.
+    #
     # Uncomment following if you want to access the admin
     "django.contrib.auth.backends.ModelBackend",
 ]
